Close and recreate JRI engine after each script eval with forked tasks
When a task is forked (created in a new processus), the forker is waiting for
the termination of the task. Since the JRI engine was created but not closed
once the script evaluation has completed, the task was hanging.

The idea of this patch is to close and recreate a JRI engine instance for
each script evaluation when forked tasks are used. If tasks are not forked, then
the same behaviour as before is applied: JRI engine instance is cached for
avoiding an expensive load operation when the context is big (lot of variables,
large script, etc.)

This patch fixes #70.