#85 Exception when using multiple R scripts inside a Task
Changed the JRI script engine loading and closing mechanism:
- the script engine is a singleton instance.
- the engine is not closed at the end of a script evaluation
- the engine is closed inside a shutdownhook.

This way, in case of a forked task, the engine will be closed before the JVM is terminated, allowing the JVM to shutdown properly.

In cas of a non-forked task, the engine will remain open across all tasks execution and will be closed on node shutdown.

The RServe behavior was not changed for now, as the engine lifecycle is much more complex.